Add package dependency [DarkMode]
This commit provides the project with swift package [DarkMode].

Swift package [DarkMode] was added as a dependency to replace existing code—the package contains the code taken from this discovery project and now is used as a third-party code.

ADDED: Dark Mode swift package instead of the existing Dark Mode source code.