Create logic for Dark Mode in Settings
This commit provides the business logic for Dark Mode option placed in Settings bundle.

Logic is very easy to get—every time when the app becomes active, the app makes a comparing of the Dark Mode value in Settings and the Dark Mode in Appearance service.

So, in case if values are different, the app overrides the Dark Mode value of Appearance service on the value from Settings, and finally, after that, the app updates its appearance in according to the new Style.

ADDED: Business logic to Dark Mode option in Settings bundle.